+++ projects/smart-table/src/lib/smart-table/smart-table.types.ts
@@ -66,7 +66,7 @@ export interface SmartTableColumnConfig {
   type: SmartTableColumnType;
   classList?: string[];
   sortPath: string;
-  canHide: boolean;
+  canHide?: boolean;
 }
 
 export enum SmartTableColumnType {