+++ projects/smart-table/src/lib/smart-table/smart-table.types.ts
@@ -1,6 +1,5 @@
 import {Filter} from '@acpaas-ui/ngx-utils';
-import {OrderBy, TableColumnFormat} from '@acpaas-ui/ngx-table';
-import {Component} from '@angular/core';
+import {ConstructableCell, OrderBy, TableColumnFormat} from '@acpaas-ui/ngx-table';
 import {LocalStorageType} from '@acpaas-ui/ngx-localstorage';
 
 export interface IModuleConfig {
@@ -79,7 +78,7 @@ export enum SmartTableColumnType {
 export interface SmartTableColumnCustomType {
   name: string;
   format?: TableColumnFormat;
-  component?: Component;
+  component?: ConstructableCell;
 }
 
 export enum SmartTableFilterType {