+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - When applying a filter, the pagination will go back to page 1
 
 ### Added
+- Option in forRoot() module method to configure custom filters
 - The search filter component as optional filter. (type: 'search-filter')
 
 ## [1.5.7] - 2020-08-21