+++ CHANGELOG.md
@@ -11,6 +11,9 @@ All notable changes to this project will be documented in this file.
 - Option in forRoot() module method to configure custom filters
 - The search filter component as optional filter. (type: 'search-filter')
 
+### Fixed
+- Display time for DateTime type columns
+
 ## [1.5.7] - 2020-08-21
 
 ### Fixed